+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Code API now generated with typedoc
+- Added typedoc generation to GitHub action
+- Added error response customisation
+
+### Fixed
+- Error handling. No more connection shutdown without trying to send a proper error response.
+- Various comments to fit the new documentation generation tool.
+- README nestjs example
+
 ## 3.0.3 - 2023-03-17
 ### Fixed
 - Install section in documentation (now use the npm package instead of GitHub repository)